backend/requirements.txt
\
fastapi
uvicorn[standard]
python-multipart
supabase
Pillow
pytesseract
python-dotenv
opencv-python
pandas
gunicorn 